BUILD_NUMBER: 527 generate the web client code  after idl change

+++ java_client/com/zhijiejiaoyu/glory_api/task/TeacherEvaluateStudentTaskRequestOrBuilder.java
@@ -65,8 +65,8 @@ public interface TeacherEvaluateStudentTaskRequestOrBuilder extends
       getFailedReasonBytes();
 
   /**
-   * <code>int64 evaluate_score = 5;</code>
+   * <code>int32 evaluate_score = 5;</code>
    * @return The evaluateScore.
    */
-  long getEvaluateScore();
+  int getEvaluateScore();
 }
